requirements.txt
QuantLib==1.32
numpy==1.26.4
pandas==2.2.2
scipy==1.13.1
yfinance==0.2.44
streamlit==1.38.0
pytest==8.3.2
pytest-xdist==3.6.1
black==24.8.0
isort==5.13.2
flake8==7.1.1
PyYAML==6.0.2
plotly==5.22.0
hypothesis==6.112.1
